+++ apps/docs/src/.vitepress/components/OnyxTypography.vue
@@ -1,7 +1,9 @@
 <script lang="ts" setup>
-import { ref } from "vue";
-import OnyxHeadline from "../../../../../packages/sit-onyx/src/components/OnyxHeadline/OnyxHeadline.vue";
-import type { HeadlineType } from "../../../../../packages/sit-onyx/src/components/OnyxHeadline/types";
+import { computed, ref } from "vue";
+import OnyxHeadline from "~components/OnyxHeadline/OnyxHeadline.vue";
+import type { HeadlineType } from "~components/OnyxHeadline/types";
+import OnyxLink from "~components/OnyxLink/OnyxLink.vue";
+import type { TextSize } from "../../../../../packages/sit-onyx/src/types/fonts";
 import DesignToken from "./DesignToken.vue";
 import DesignTokenCard from "./DesignTokenCard.vue";
 import DesignTokenHeader from "./DesignTokenHeader.vue";
@@ -14,8 +16,8 @@ export type TypographyToken = {
   name: string;
   /** HTML element to render. */
   htmlTag: HeadlineType | "p" | "a";
-  /** Font size to use if `htmlTag` is <p> or <a> */
-  fontSize?: "big" | "default" | "small";
+  /** Text size to use if `htmlTag` is <p> or <a> */
+  textSize?: TextSize;
 };
 
 const AVAILABLE_FONT_TABS = ["Source Sans 3", "Source Code Pro"] as const;
@@ -29,16 +31,20 @@ const props = defineProps<{
 
 const previewText = "onyx design system" as const;
 const currentTab = ref<AvailableFontTab>(AVAILABLE_FONT_TABS[0]);
+
+const isMonospace = computed(() => currentTab.value === "Source Code Pro");
+
+const getTextSizeClass = (fontSize?: TextSize) => {
+  if (!fontSize || fontSize === "default") return "";
+  return `onyx-text--${fontSize}`;
+};
 </script>
 
 <template>
   <section class="typography">
     <DesignTokenHeader v-model="currentTab" :tabs="AVAILABLE_FONT_TABS" />
 
-    <div
-      class="typography__content"
-      :class="{ 'typography__content--mono': currentTab === 'Source Code Pro' }"
-    >
+    <div class="typography__content" :class="isMonospace ? 'onyx-text--monospace' : ''">
       <DesignTokenCard
         v-for="token in props.tokens"
         :key="token.name"
@@ -46,18 +52,24 @@ const currentTab = ref<AvailableFontTab>(AVAILABLE_FONT_TABS[0]);
         :wide-name="props.wideName"
       >
         <template #name>
-          <p v-if="token.htmlTag === 'p'" :class="`font-size--${token.fontSize ?? 'default'}`">
-            {{ previewText }}
-          </p>
-          <a
-            v-else-if="token.htmlTag === 'a'"
-            :class="`font-size--${token.fontSize ?? 'default'}`"
-            href="#"
+          <p
+            v-if="token.htmlTag === 'p'"
+            class="onyx-text"
+            :class="getTextSizeClass(token.textSize)"
           >
             {{ previewText }}
-          </a>
+          </p>
 
-          <OnyxHeadline v-else :is="token.htmlTag" :monospace="currentTab === 'Source Code Pro'">
+          <OnyxLink
+            v-else-if="token.htmlTag === 'a'"
+            href="#"
+            class="onyx-text"
+            :class="getTextSizeClass(token.textSize)"
+          >
+            {{ previewText }}
+          </OnyxLink>
+
+          <OnyxHeadline v-else :is="token.htmlTag" :monospace="isMonospace">
             {{ previewText }}
           </OnyxHeadline>
         </template>
@@ -88,35 +100,11 @@ const currentTab = ref<AvailableFontTab>(AVAILABLE_FONT_TABS[0]);
   }
 }
 
-.font-size {
-  &--big {
-    font-size: 1.25rem;
-    line-height: 1.75rem;
-    font-weight: 400;
-  }
-
-  &--default {
-    font-size: 1rem;
-    line-height: 1.5rem;
-    font-weight: 400;
-  }
-
-  &--small {
-    font-size: 0.8125rem;
-    line-height: 1.25rem;
-    font-weight: 400;
-  }
-}
-
 .typography {
   &__content {
     display: flex;
     flex-direction: column;
     gap: var(--onyx-spacing-lg);
-
-    &--mono {
-      font-family: var(--onyx-font-family-mono);
-    }
   }
 }
 </style>