+++ packages/sit-onyx/src/utils/storybook.ts
@@ -1,3 +1,6 @@
+import { sourceCodeTransformer } from "@sit-onyx/storybook-utils";
+import type { Decorator, StoryContext } from "@storybook/vue3";
+
 /**
  * Defines the control for a Storybook argType to be a select/dropdown of
  * all available onyx icons.
@@ -38,3 +41,34 @@ export const defineIconSelectArgType = () => {
     },
   };
 };
+
+export const createIconSourceCodeTransformer = (propertyName: string) => {
+  const iconArgType = defineIconSelectArgType();
+
+  return (sourceCode: string, ctx: Pick<StoryContext, "args">) => {
+    // using this custom transformer would override the default one
+    // so we are calling the default transformer here
+    const code = sourceCodeTransformer(sourceCode);
+    if (!ctx.args[propertyName] || typeof ctx.args[propertyName] !== "string") return code;
+
+    const iconName = iconArgType.control.labels[ctx.args[propertyName] as string];
+
+    return `<script lang="ts" setup>
+import icon from "@sit-onyx/icons/${iconName}.svg?raw";
+</script>
+${code.replace(new RegExp(` ${propertyName}=['"].*['"]`), ` :${propertyName}="icon"`)}`;
+  };
+};
+
+/**
+ * Storybook decorator that wraps the story with a <div> that sets the text color
+ * to neutral intense.
+ * Useful if the component uses "currentColor" in its CSS.
+ */
+export const textColorDecorator: Decorator = (story) => ({
+  components: { story },
+  template: `
+  <div style="color: var(--onyx-color-text-icons-neutral-intense)">
+    <story />
+  </div>`,
+});