+++ packages/storybook-utils/vitest.config.ts
@@ -1,0 +1,12 @@
+import { defineConfig } from "vitest/config";
+
+export default defineConfig({
+  test: {
+    environment: "jsdom",
+    include: ["src/**/*.spec.ts"],
+    coverage: {
+      include: ["src"],
+      provider: "v8",
+    },
+  },
+});