+++ packages/sit-onyx/CHANGELOG.md
@@ -1,5 +1,23 @@
 # sit-onyx
 
+## 1.0.0-alpha.22
+
+### Minor Changes
+
+- f859db6: feat(OnyxLoadingIndicator): animate dots
+
+## 1.0.0-alpha.21
+
+### Minor Changes
+
+- 6324d32: feat: add `OnyxLink` component
+
+## 1.0.0-alpha.20
+
+### Minor Changes
+
+- 6045208: feat: add OnyxLoading indicator (circle)
+
 ## 1.0.0-alpha.19
 
 ### Minor Changes